moved users on join
